Segregate Database Provider Interface
