+++ Poc.TextProcessor.ResourceAccess.Database/Abstractions/IDatabaseWriterProvider.cs
@@ -1,0 +1,8 @@
+﻿namespace Poc.TextProcessor.ResourceAccess.Database.Abstractions
+{
+    public interface IDatabaseWriterProvider
+    {
+        T Save<T>(T entity) where T : class;
+        Task<T> SaveAsync<T>(T entity) where T : class;
+    }
+}
