+++ Poc.TextProcessor.ResourceAccess.Repositories/TextSortRepository.cs
@@ -7,7 +7,7 @@ using Poc.TextProcessor.ResourceAccess.Repositories.Base;
 
 namespace Poc.TextProcessor.ResourceAccess.Repositories
 {
-    public class TextSortRepository(IDatabaseProvider databaseProvider) : RepositoryBase(databaseProvider), ITextSortRepository
+    public class TextSortRepository(IDatabaseReaderProvider databaseProvider) : RepositoryReaderBase(databaseProvider), ITextSortRepository
     {
         public IEnumerable<TextSort> List()
         {
